requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
biopython>=1.83
rdkit>=2023.9.5
py3Dmol>=2.1.0
stmol>=0.0.9
requests>=2.31.0
altair>=5.2.0
openai>=1.10.0